pyproject.toml
[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
^/(
  | packages
)/
'''
force-exclude = 'lib/galaxy/util/jstree.py'

[tool.darker]
isort = true

[tool.poetry]
name = "galaxy"
version = "22.01.dev0"
description = "Galaxy is an open, web-based platform for accessible, reproducible, and transparent computational research."
authors = ["Galaxy Contributors"]
license = "AFL-3.0"
readme = "README.rst"
homepage = "https://galaxyproject.org/"
repository = "https://github.com/galaxyproject/galaxy/"
documentation = "https://docs.galaxyproject.org/"

[[tool.poetry.source]]
name = "galaxyproject"
url = "https://wheels.galaxyproject.org/simple"

[tool.poetry.dependencies]
a2wsgi = "*"
aiofiles = "*"
alembic = "*"
apispec = "*"
Babel = "*"
bdbag = ">=1.6.3"
Beaker = "1.11.0"
bioblend = "*"
bleach = "*"
boltons = "*"
boto = "*"
bx-python = "*"
celery = "*"
Cheetah3 = "!=3.2.6.post2"  # yanked release, https://github.com/python-poetry/poetry/issues/2453
cloudauthz = "==0.6.0"
cloudbridge = "*"
circus = "*"
cwltool = "==3.1.20211107152837"
dictobj = "*"
docutils = "!=0.17, !=0.17.1"
edam-ontology = "*"
fastapi = ">=0.68.2, !=0.69.0, !=0.70.0, !=0.70.1"  # https://github.com/tiangolo/fastapi/issues/4041
fastapi-utils = "*"
fs = "*"
future = "*"
galaxy_sequence_utils = "*"
gravity = ">=0.10.0"
gunicorn = "*"
gxformat2 = "*"
h5py = "*"
importlib_resources = "*"
isa-rwval = "*"
kombu = "*"
lagom = "*"
Mako = "*"
Markdown = "*"
MarkupSafe = "*"
Mercurial = "*"
mrcfile = "*"
msal = "*"
nodeenv = "*"
nose = "*"
numpy = "*"
paramiko = ">=2.12.0"
Parsley = "*"
Paste = "*"
pebble = "*"
psutil = "*"
pulsar-galaxy-lib = ">=0.14.13"
pycryptodome = "*"
pydantic = "*"
PyJWT = "*"
pykwalify = "*"
pylibmagic = "*"
pyparsing = "*"
pysam = "*"
python = ">=3.7,<3.11"  # Keep maximum version strict to appease numpy and scipy
python-dateutil = "*"
python-magic = "*"
python-multipart = "*"  # required to support form parsing in FastAPI/Starlette
PyYAML = "*"
refgenconf = ">=0.12.0"
requests = "*"
Routes = "*"
social-auth-core = {version = "==4.0.3", extras = ["openidconnect"]}
sortedcontainers = "*"
SQLAlchemy = ">=1.4.25,<2"
sqlalchemy-migrate = "*"  # required to support tool shed
sqlitedict = "*"
sqlparse = "*"
fs-sshfs = "*"
starlette = "*"
starlette-context = "*"
svgwrite = "*"
tifffile = "*"
tuswsgi = "*"
typing-extensions = "*"
uvicorn = "*"
uvloop = "*"
WebOb = "*"
Whoosh = "*"
zipstream-new = "*"
rucio-clients = "^1.29.10"

[tool.poetry.dev-dependencies]
ase = ">=3.18.1"
black = "^22.1.0"
cwltest = "2.2.20210901154959"
darker = "*"
fluent-logger = "*"
httpx = "*"
isort = "^5.10.1"
lxml = "!=4.2.2"
markdown-it-reporter = "*"
NoseHTML = "*"
PyGithub = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-celery = "*"
pytest-cov = "*"
pytest-html = "*"
python-irodsclient = "!=1.1.2"  # https://github.com/irods/python-irodsclient/issues/356
pytest-json-report = "*"
pytest-mock = "*"
pytest-postgresql = "!=3.0.0"  # https://github.com/ClearcodeHQ/pytest-postgresql/issues/426
pytest-shard = "*"
recommonmark = "*"
responses = "*"
selenium = "*"
Sphinx = ">=4.0"  # support docutils 0.17
sphinx_markdown_tables = "*"
sphinx_rtd_theme = ">=1.0.0"
statsd = "*"
testfixtures = "*"
tuspy = "*"
twill = "*"
watchdog = "*"
